CP-27474: prepare 1.0.3 release (#190)
* CP-27062: don't delete jobs after they have finished

Some tools for managing Kubernetes, notably Flux, don't play nicely
with jobs that automatically clean themselves up by setting a
`ttlSecondsAfterFinished`; they see that the job has been deleted,
and recreate the job.  This results in an endless cycle of the job
being run then deleting itself over and over, when we really only
want it to be run once.

The reason we were using `ttlSecondsAfterFinished` was because we *do*
want the job to run when upgrading the chart.  This change gives us
that benefit without relying on ttlSecondsAfetrFinished.

Specifically, we now append a (partial) SHA-256 checksum of the chart
version and the configuration values in the job name. This means that,
from Kubernetes' perspective, whenever a chart is upgraded (even if it
is just a configuration change), a completely new job is created and
the old one is destroyed.

* CP-27063: have insights-controller send data to Gator when appropriate

* additional changes for 0.2.1 compatability

* update metrics destination template to format correctly for 0.2.1

* revert to original url definition

---------

Co-authored-by: Evan Nemerson <evan.nemerson@cloudzero.com>