CP-29248: clean up and document Prometheus config (#247)
* Indentation changes in Prometheus configuration.

These shouldn't actually change anything; the extra indentation just
makes it easier to read documentation comments attached to the items,
which I'll add in a follow-up commit. This just creates a lot of
noise so I wanted to isolate it from the rest of the changes.

* Add documentation to scrape config.

* Remove unnecessary properties from Prometheus configuration.

These all just set the value to the same value as the default, which
just clutters up the configuration.

The only exception is the scrape_protocols, where the default also
contains `PrometheusProto`, which is the newest (and supposedly most
efficient) protocol. Based on internal conversations, it sounds like
the initial code for these configurations was copied from an example
which was likely written before PrometheusProto existed.