CP-29254: add default PodDisruptionBudget (#252)
* CP-29254: add a default Pod Disruption Budget, and use it

Set the minimum number of available pods to 1 for our deployments
should be Kubernetes much less prone to preepmt our pods, which is
important for data quality.

* Use jobConfigID when generating validator-job checksum

This just allow us to reduce noise when regenerating the Helm test
templates (tests/helm/template/*.yml).

* Don't clobber defaults when merging anti-affinity in webhook

Previously, we merged $podAntiAffinity into
.Values.insightsController.server.affinity, which isn't what we want
because it was adding the anti-affinity everywhere; i.e., clobbering
the default values.

This will create a copy of the defaults and merge into that instead,
leaving the defaults alone.