CP-29492: clean up generation of annotation checksums (#260)
Previously, we didn't always allow jobConfigID to override the checksum
when adding it to annotations. This resulted in some unnecessary noise
sometimes when regenerating the manifests.

I've also changed the value we use use in the overrides files, of both
jobConfigID and apiKey, to make it clear that they are not, in fact,
secrets that we're leaking, but just dummy values.

Finally, instead in a few places we were merging stuff into the default
annotations, since the merge function in Helm will actually merge into
the destination instead of merging everything into a new map. Using a
deepCopy of the destination instead takes care of this. The effect is
that in many places we will no longer be including annotations we didn't
intend to.