CP-30255: periodically close connections to Webhook Server (#334)
We're a bit worried about clients creating a sticky connnection to
the Webhook Server once, and subsequent requests not being load-
balanced because they are staying connected to the same pod.

This just copies a trick from the collector, which is to randomly
send "Connection: close" headers in order to cause the client to
reconnect, at which point they will go through the round-robin for
the webhook service again.