CP-29944: increase the costMaxInterval from 10m to 30m (#351)
This value tells the collector the maximum amount of time to buffer
metrics. While ideally we'd like to wait until reaching `maxRecords`
metrics (currently 1.5m), we're seeing lots of small clusters which
aren't getting anywhere near that number of rows. Reducing this
timeout to 30m effectively reduces the number of files sent by such
clusters to 1/3 of its previous value.

This also decreases the timout for sending observability files to
10m, in order to make sure the cluster is still communicating often
enough to be shown as online in the UI.